Add vagrant setup
Setting up an environment with the dependencies needed to build and test
Ruby is daunting for a newcomer.  Vagrant provides an automated way to
set up an environment based on a virtual machine.  This provides for a
consistent experience and fast onboarding of new MRI Ruby contributors.

To use the vagrant setup, VirtualBox and Vagrant will need to be
installed on the development machine.  In the terminal, change to the
directory containing the Ruby repository and run the following command.

  $ vagrant up

This will install and setup a virtual machine ready for contributing to
MRI Ruby.  Follow the directions at the end of the installation to get
started.